Handle file_info protocal

+++ projects/piretoutpire/src/file/file_chunk.rs
@@ -1,4 +1,4 @@
-use crate::utils::div::div_ceil;
+use crate::utils::div_ceil;
 use errors::{bail, AnyResult, Context};
 use std::{
     cmp::min,
@@ -7,18 +7,18 @@ use std::{
     path::Path,
 };
 
-// pub const DEFAULT_CHUNK_SIZE: u64 = 1024 * 1024; // 1 Mo chunks.
-pub const DEFAULT_CHUNK_SIZE: u64 = 2; // 1 Mo chunks.
+// pub const DEFAULT_CHUNK_SIZE: u32 = 1024 * 1024; // 1 Mo chunks.
+pub const DEFAULT_CHUNK_SIZE: u32 = 2; // 1 Mo chunks.
 
 pub type FileChunk = FileFixedSizedChunk<DEFAULT_CHUNK_SIZE>;
 
-pub struct FileFixedSizedChunk<const CHUNK_SIZE: u64> {
-    file_size: u64,
+pub struct FileFixedSizedChunk<const CHUNK_SIZE: u32> {
+    file_size: u32,
     writer: BufWriter<File>,
     reader: BufReader<File>,
 }
 
-impl<const CHUNK_SIZE: u64> FileFixedSizedChunk<CHUNK_SIZE> {
+impl<const CHUNK_SIZE: u32> FileFixedSizedChunk<CHUNK_SIZE> {
     // Open an already existing file. File must exists.
     pub fn open_existing<P>(path: P) -> AnyResult<Self>
     where
@@ -26,14 +26,14 @@ impl<const CHUNK_SIZE: u64> FileFixedSizedChunk<CHUNK_SIZE> {
     {
         let file = OpenOptions::new().read(true).open(&path)?;
         Ok(Self {
-            file_size: file.metadata().context("can't get file size")?.len(),
+            file_size: file.metadata().context("can't get file size")?.len() as u32,
             writer: BufWriter::new(OpenOptions::new().write(true).open(&path)?),
             reader: BufReader::new(file),
         })
     }
 
     // Open and allocate space for a new file. Existing file will be truncated.
-    pub fn open_new<P>(path: P, preallocated_size: u64) -> AnyResult<Self>
+    pub fn open_new<P>(path: P, preallocated_size: u32) -> AnyResult<Self>
     where
         P: AsRef<Path>,
     {
@@ -55,7 +55,7 @@ impl<const CHUNK_SIZE: u64> FileFixedSizedChunk<CHUNK_SIZE> {
     }
 
     // Read a chunk by its index.
-    pub fn read_chunk(&mut self, chunk_id: u64) -> AnyResult<Vec<u8>> {
+    pub fn read_chunk(&mut self, chunk_id: u32) -> AnyResult<Vec<u8>> {
         read_range(
             &mut self.reader,
             chunk_id * CHUNK_SIZE,
@@ -64,7 +64,7 @@ impl<const CHUNK_SIZE: u64> FileFixedSizedChunk<CHUNK_SIZE> {
     }
 
     // Write a chunk to its index.
-    pub fn write_chunk(&mut self, chunk_id: u64, data: &[u8]) -> AnyResult<()> {
+    pub fn write_chunk(&mut self, chunk_id: u32, data: &[u8]) -> AnyResult<()> {
         write_range(
             &mut self.writer,
             chunk_id * CHUNK_SIZE,
@@ -74,18 +74,18 @@ impl<const CHUNK_SIZE: u64> FileFixedSizedChunk<CHUNK_SIZE> {
     }
 
     // Number of chunks the file is currently split in.
-    pub fn nb_chunks(&self) -> u64 {
+    pub fn nb_chunks(&self) -> u32 {
         div_ceil(self.file_size, CHUNK_SIZE)
     }
 
     // Size of the file when retrieved.
-    pub fn file_size(&self) -> u64 {
+    pub fn file_size(&self) -> u32 {
         self.file_size
     }
 }
 
 // Read a range of bytes from a buffer.
-fn read_range<T>(br: &mut T, from: u64, to: u64) -> AnyResult<Vec<u8>>
+fn read_range<T>(br: &mut T, from: u32, to: u32) -> AnyResult<Vec<u8>>
 where
     T: BufRead + Seek,
 {
@@ -100,14 +100,14 @@ where
 }
 
 // Write a range of bytes into a buffer.
-fn write_range<T>(bw: &mut T, from: u64, to: u64, data: &[u8]) -> AnyResult<()>
+fn write_range<T>(bw: &mut T, from: u32, to: u32, data: &[u8]) -> AnyResult<()>
 where
     T: Write + Seek,
 {
     if from > to {
         bail!("From({}) is > than to({}) with data({})", from, to, data.len());
     }
-    if to - from != data.len() as u64 {
+    if to - from != data.len() as u32 {
         bail!(
             "Invalid chunk size from({}) + to({}) != data({})",
             from,
